feat: rearrange functions to files
- private functions to helpers
- public functions to gobrew
- use os variables only on main function
- setup gobrew with one function
